Add teacherId prefix only if it does not exist in iamService. Move iamService to service package.
