+++ src/main/java/fi/helsinki/moodi/service/synchronize/process/SynchronizingProcessor.java
@@ -18,7 +18,7 @@
 package fi.helsinki.moodi.service.synchronize.process;
 
 import fi.helsinki.moodi.exception.ProcessingException;
-import fi.helsinki.moodi.integration.iam.IAMService;
+import fi.helsinki.moodi.service.iam.IAMService;
 import fi.helsinki.moodi.integration.moodle.MoodleEnrollment;
 import fi.helsinki.moodi.integration.moodle.MoodleService;
 import fi.helsinki.moodi.integration.moodle.MoodleUser;
